Create AICivilian Tree with Escape and Roam Task
Create AICivilian Behavior Tree. With two branches managed by a Selector Node. In the first branch (More priority) theres a Sequence Node to check if the player is in the AI FOV range and executes the run away task. The other branch is just a task Node to roam around the map